move function params
