+++ services/autorust/codegen/src/codegen_operations/function_params.rs
@@ -1,5 +1,6 @@
 use autorust_openapi::{CollectionFormat, ParameterIn};
-use proc_macro2::Ident;
+use proc_macro2::{Ident, TokenStream};
+use quote::{quote, ToTokens};
 
 use crate::codegen::{parse_query_params, TypeNameCode};
 use crate::identifier::SnakeCaseIdent;
@@ -124,5 +125,26 @@ impl From<&ParameterIn> for ParamKind {
             ParameterIn::Body => Self::Body,
             ParameterIn::FormData => Self::FormData,
         }
+    }
+}
+
+#[derive(Clone)]
+pub struct FunctionCallParamsCode(pub FunctionParams);
+
+impl ToTokens for FunctionCallParamsCode {
+    fn to_tokens(&self, tokens: &mut TokenStream) {
+        let mut params: Vec<TokenStream> = Vec::new();
+        for FunctionParam {
+            variable_name, type_name, ..
+        } in self.0.required_params()
+        {
+            let mut type_name = type_name.clone();
+            let is_vec = type_name.is_vec();
+            type_name.impl_into(!is_vec);
+            params.push(quote! { #variable_name: #type_name });
+        }
+        let slf = quote! { &self };
+        params.insert(0, slf);
+        tokens.extend(quote! { #(#params),* })
     }
 }
